:memo: :: edit example -> school_of_company
